Fix: 0163 failed to apply — carried_count changed type
Migration 0163 was rejected by Postgres with 42P16, "cannot change
data type of view column carried_count from bigint to integer".

CREATE OR REPLACE VIEW can add trailing columns but cannot rename,
reorder or retype an existing one. The original expression was
coalesce(sum(...), 0), which is bigint; I replaced it with a bare 0,
which is integer. Now cast explicitly with 0::bigint.

Dropping and recreating instead is not an option worth reaching for:
annual_goal_progress depends on this view, so DROP would need CASCADE
and would take the rollups with it.

Adds two guards that catch this class without needing a database,
since there is no local Postgres here to execute against:
- the projected column list of the latest definition must match the
  original, in order
- carried_count must carry the ::bigint cast

Everything else in 0163 is unchanged and was already correct: the
remaining columns keep their original expressions and therefore their
original types.

+++ supabase/migrations/0163_fix_priority_progress_statuses.sql
@@ -50,7 +50,16 @@ select
   coalesce(sum(case when c.status = 'missed' then 1 else 0 end), 0) as missed_count,
   -- Vestigial: 'carried' was dropped in migration 0011. Kept in the
   -- projection so the row shape stays stable for existing callers.
-  0               as carried_count,
+  --
+  -- The ::bigint cast is REQUIRED, not decoration. CREATE OR REPLACE
+  -- VIEW cannot change an existing column's data type, and the
+  -- original expression was coalesce(sum(...), 0), which is bigint.
+  -- A bare 0 is integer, and Postgres rejects the whole statement
+  -- with 42P16 "cannot change data type of view column". Dropping and
+  -- recreating instead is not an option: annual_goal_progress depends
+  -- on this view, so DROP would need CASCADE and would take the
+  -- rollups with it.
+  0::bigint       as carried_count,
   count(c.id)     as denominator,
   case
     when p.status = 'complete' then 100
